Final round of ip_bot fixes (hopefully)

+++ ip_bot.py
@@ -35,20 +35,20 @@ class PollIPCog(commands.Cog):
             await bot.wait_until_ready()
             session = Session()
             global ip_addr
-            # if socket.gethostbyname(hostname) != ip_addr:
             s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
             s.connect(("8.8.8.8", 80))
-            ip_addr = s.getsockname()[0]
-            channels = []
-            for channel in session.query(Channel).all():
-                c = bot.get_channel(channel.channel_id)
-                if c:
-                    channels.append(c)
-                else:
-                    session.delete(channel)
-            message = f"My new IP addr is {ip_addr}" 
-            for channel in channels:
-                await channel.send(message)
+            if s.getsockname()[0] != ip_addr:
+                ip_addr = s.getsockname()[0]
+                channels = []
+                for channel in session.query(Channel).all():
+                    c = bot.get_channel(channel.channel_id)
+                    if c:
+                        channels.append(c)
+                    else:
+                        session.delete(channel)
+                message = f"My new IP addr is {ip_addr}"
+                for channel in channels:
+                    await channel.send(message)
         except Exception:
             traceback.print_exc()
             session.rollback()
